+++ CHANGELOG.md
@@ -176,3 +176,4 @@
 ### (2015/5/9)
 - Send paths to nata-server
 - Show act paths in the apk view and can replay act paths
+- Rewrite controllers to decouple web_router